requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
configparser==5.0.2
Cython==0.29.24
idna==3.2
numpy==1.21.1
pandas==1.3.1
pyaes==1.6.1
pyasn1==0.4.8
pyTelegramBotAPI==3.8.2
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
schedule==1.1.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
Telethon==1.23.0
tornado==6.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
backports.entry-points-selectable==1.1.0
distlib==0.3.2
filelock==3.0.12
platformdirs==2.2.0
virtualenv==20.7.2